requirements.txt
pandas
scikit-learn
matplotlib
seaborn
streamlit
opencv-python
numpy
Pillow
kivy
kivy-garden
plyer
fastapi
uvicorn
python-multipart
face-recognition